requirements.txt
keras
tensorflow
pandas
sklearn
newsapi-python
git+https://github.com/m-wrzr/populartimes.git
